Add an example with lots of markers
