+++ example/lib/clustering_many_markers_page.dart
@@ -1,0 +1,106 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map_marker_cluster/flutter_map_marker_cluster.dart';
+import 'package:flutter_map_marker_cluster_example/drawer.dart';
+import 'package:latlong2/latlong.dart';
+
+class ClusteringManyMarkersPage extends StatefulWidget {
+  static const String route = 'clusteringManyMarkersPage';
+
+  @override
+  _ClusteringManyMarkersPageState createState() =>
+      _ClusteringManyMarkersPageState();
+}
+
+class _ClusteringManyMarkersPageState extends State<ClusteringManyMarkersPage> {
+  static const totalMarkers = 2000.0;
+  final minLatLng = LatLng(49.8566, 1.3522);
+  final maxLatLng = LatLng(58.3498, -10.2603);
+
+  late List<Marker> markers;
+
+  @override
+  void initState() {
+    final latitudeRange = maxLatLng.latitude - minLatLng.latitude;
+    final longitudeRange = maxLatLng.longitude - minLatLng.longitude;
+
+    final stepsInEachDirection = sqrt(totalMarkers).floor();
+    final latStep = latitudeRange / stepsInEachDirection;
+    final lonStep = longitudeRange / stepsInEachDirection;
+
+    markers = [];
+    for (var i = 0; i < stepsInEachDirection; i++) {
+      for (var j = 0; j < stepsInEachDirection; j++) {
+        final latLng = LatLng(
+          minLatLng.latitude + i * latStep,
+          minLatLng.longitude + j * lonStep,
+        );
+
+        markers.add(
+          Marker(
+            height: 30,
+            width: 30,
+            point: latLng,
+            builder: (ctx) => Icon(Icons.pin_drop),
+          ),
+        );
+      }
+    }
+
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('Clustering Many Markers Page')),
+      drawer: buildDrawer(context, ClusteringManyMarkersPage.route),
+      body: FlutterMap(
+        options: MapOptions(
+          center: LatLng((maxLatLng.latitude + minLatLng.latitude) / 2,
+              (maxLatLng.longitude + minLatLng.longitude) / 2),
+          zoom: 6,
+          maxZoom: 15,
+          plugins: [
+            MarkerClusterPlugin(),
+          ],
+        ),
+        children: <Widget>[
+          TileLayerWidget(
+            options: TileLayerOptions(
+              urlTemplate: 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+              subdomains: ['a', 'b', 'c'],
+            ),
+          ),
+          MarkerClusterLayerWidget(
+            options: MarkerClusterLayerOptions(
+              maxClusterRadius: 45,
+              size: Size(40, 40),
+              anchor: AnchorPos.align(AnchorAlign.center),
+              fitBoundsOptions: FitBoundsOptions(
+                padding: EdgeInsets.all(50),
+                maxZoom: 15,
+              ),
+              markers: markers,
+              builder: (context, markers) {
+                return Container(
+                  decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(20.0),
+                      color: Colors.blue),
+                  child: Center(
+                    child: Text(
+                      markers.length.toString(),
+                      style: TextStyle(color: Colors.white),
+                    ),
+                  ),
+                );
+              },
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
